html/semantics/embedded-content/the-audio-element/audio-loading-attr-lazy.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<audio id="target" loading="lazy" controls></audio>

<script>
  const audio = document.getElementById("target");

  test(() => {
    assert_equals(
      audio.loading,
      "lazy",
      "Audio loading attribute should be 'lazy' when set to 'lazy'"
    );
  }, "Audio loading attribute should reflect the loading content attribute");
</script>
